Fix vim backups and clean aliases
Vim, by default, uses the current file as a backup and creates
a new file that it edits.  When saving, it moves the new file
to the original name.  This has the impact of changing the inode
of the file.  When mounting the file directly into docker, the
inode is what is mounted, so changing the inode makes any changes
done by vim not visible to docker.

Using
 set backupcopy=yes
tells vim to use a different file as a backup and make its edit
into the original file, thus keeping the same inode.